Add files via upload
The code shows how powerful macros are. One cautionary tale, nowadays in the C and C++ community, there's much less reliance on macros especially ones that involve code. In general for code segments, we either let the compiler optimize it or for example in the C++ language and some C compilers you can call a function inline and that asks the compiler to optimize it much like a would be for macro and that is much safer code, more check